LineNumber Import format helper class

+++ formats/utils/KeyValueReader.h
@@ -28,7 +28,7 @@ namespace displace {
                 std::string get(const std::string &key, std::string value = std::string()) const;
                 bool isSet(const std::string &key) const;
 
-                int numKeyValues() const {
+                unsigned long numKeyValues() const {
                     return mConfig.size();
                 }
             private:
